+++ src/App.js
@@ -1,12 +1,13 @@
 import './App.css';
-import { Checkbox, FormGroup, FormControlLabel, FormLabel, Grid, Paper, Box, Button, Typography, Popover, List, ListItem } from '@mui/material';
+import { Checkbox, FormGroup, FormControlLabel, FormLabel, Grid, Paper, Box, Button } from '@mui/material';
 import materias_plan86 from "./plan_86.json";
 import materias_plan23 from "./plan_23.json";
 import { useEffect, useState } from 'react';
 import { useMaterias86 } from './utils/useMaterias86';
 import DeleteIcon from '@mui/icons-material/Delete';
 import SnowboardingIcon from '@mui/icons-material/Snowboarding';
-import InfoIcon from '@mui/icons-material/Info';
+import Materia86 from './components/Materia86';
+import Materia23 from './components/Materia23';
 
 function App() {
   const [creditos, setCreditos] = useState(0);
@@ -14,8 +15,6 @@ function App() {
   const [creditosTransicion, setCreditosTransicion] = useState(0);
   const [materias86, setMaterias86] = useMaterias86("materias86-calculadorBilbao", []);
   const [materias23, setMaterias23] = useState([]);
-  const [popoverAnchor, setPopoverAnchor] = useState(null);
-  const [popoverMateria, setPopoverMateria] = useState("");
 
   const agregarMateria86 = (materia) => {
     if (materias86.includes(materia)) return;
@@ -45,25 +44,6 @@ function App() {
   const eliminarMateria86 = (materia) => {
     setMaterias86(materias86.filter(m => m.nombre !== materia.nombre));
   };
-
-  const handleCheck = (e, materia) => {
-    if (e.target.checked) {
-      agregarMateria86(materia);
-    } else {
-      eliminarMateria86(materia);
-    }
-  };
-
-  const handlePopoverClick = (e, materia) => {
-    setPopoverAnchor(e.currentTarget);
-    setPopoverMateria(materia.nombre);
-  };
-
-  const creditosExtraNecesarios = (equivalencia) => {
-    return 0 - equivalencia.creditos
-      - materias_plan86.obligatorias.filter(m => equivalencia.materias.includes(m.nombre)).reduce((a, b) => a.creditosExtra??0 + b.creditosExtra??0, 0)
-      - materias_plan86.electivas.filter(m => equivalencia.materias.includes(m.nombre)).reduce((a, b) => a.creditosExtra??0 + b.creditosExtra??0, 0)
-  }
 
   useEffect(() => {
     let _materias23 = [];
@@ -140,10 +120,12 @@ function App() {
             <h2>Obligatorias</h2>
             <FormGroup>
               {materias_plan86.obligatorias.map((materia, idx) =>
-                <FormControlLabel
+                <Materia86
                   key={`${materia.nombre}-86`}
-                  control={<Checkbox onChange={(e) => handleCheck(e, materia)} checked={materias86.some(m => m.nombre === materia.nombre)} />}
-                  label={materia.nombre}
+                  materia={materia}
+                  checked={materias86.some(m => m.nombre === materia.nombre)}
+                  onCheck={agregarMateria86}
+                  onUncheck={eliminarMateria86}
                 />
               )}
             </FormGroup>
@@ -154,10 +136,12 @@ function App() {
               <FormGroup key={orientacion.nombre} sx={{marginBottom: "1em"}}>
                 <FormLabel>{orientacion.nombre}</FormLabel>
                 {orientacion.materias.map((materia, idx) =>
-                  <FormControlLabel
+                  <Materia86
                     key={`${materia.nombre}-86`}
-                    control={<Checkbox onChange={(e) => handleCheck(e, materia)} checked={materias86.some(m => m.nombre === materia.nombre)}/>}
-                    label={materia.nombre}
+                    materia={materia}
+                    checked={materias86.some(m => m.nombre === materia.nombre)}
+                    onCheck={agregarMateria86}
+                    onUncheck={eliminarMateria86}
                   />
                 )}
               </FormGroup>
@@ -169,10 +153,12 @@ function App() {
             <h2>Electivas</h2>
             <FormGroup>
               {materias_plan86.electivas.map(materia =>
-                <FormControlLabel
+                <Materia86
                   key={`${materia.nombre}-86`}
-                  control={<Checkbox onChange={e => handleCheck(e, materia)} checked={materias86.some(m => m.nombre === materia.nombre)}/>}
-                  label={materia.nombre}
+                  materia={materia}
+                  checked={materias86.some(m => m.nombre === materia.nombre)}
+                  onCheck={agregarMateria86}
+                  onUncheck={eliminarMateria86}
                 />
               )}
             </FormGroup>
@@ -183,39 +169,11 @@ function App() {
             <h2>Plan 2023</h2>
             <FormGroup>
               {materias_plan23.map(materia =>
-                <div key={`${materia.nombre}-23`}>
-                  <FormControlLabel
-                    control={
-                      <Checkbox
-                        onClick={e => e.preventDefault()}
-                        checked={materias23.includes(materia.nombre)}
-                      />
-                    }
-                    disableTypography={true}
-                    label={
-                      <>
-                        <Typography>{materia.nombre}</Typography>
-                        <InfoIcon onClick={e => handlePopoverClick(e, materia)} sx={{marginLeft: "0.4em"}} />
-                      </>
-                    }
-                  />
-                  <Popover
-                    open={popoverMateria === materia.nombre}
-                    anchorEl={popoverAnchor}
-                    onClose={() => setPopoverAnchor(null) || setPopoverMateria("")}
-                    anchorOrigin={{vertical: "top", horizontal: "right"}}
-                    transformOrigin={{horizontal: "right", vertical: "top"}}
-                  >
-                    <List>
-                      {materia.equivalencias.map((equivalencia, idx) =>
-                        <ListItem key={`${materia.nombre}-23-equivalencia${idx}`} sx={{p: 2}}>
-                          {equivalencia.materias.join(' + ')}
-                          {creditosExtraNecesarios(equivalencia) > 0 && ` + ${creditosExtraNecesarios(equivalencia)} créditos`}
-                        </ListItem>
-                      )}
-                    </List>
-                  </Popover>
-                </div>
+                <Materia23
+                  key={`${materia.nombre}-23`}
+                  materia={materia}
+                  checked={materias23.includes(materia.nombre)}
+                />
               )}
               <FormControlLabel
                 control={
